Fixed netting event shortcuts
These functions:

- get_netting_channel_closed_events
- get_netting_channel_deposit_events
- get_netting_channel_settled_events

Weren't using the `netting_channel_identifier` argument, therefor the
events for all channels were polled. This was okay since the functions
were used only for testing.

This moves the functions to the testing module and fix the
implementation to use the parameter.
